+++ worldwind/src/commonMain/kotlin/earth/worldwind/render/RenderContext.kt
@@ -65,7 +65,7 @@ open class RenderContext {
     private var pixelSizeFactor = 0.0
     private val userProperties = mutableMapOf<Any, Any>()
     val drawablePools = mutableMapOf<Any, Pool<*>>()
-    private val textRenderer = TextRenderer(this)
+    private var textRenderer = TextRenderer(this)
     private val scratchTextCacheKey = TextCacheKey()
     private val scratchVector = Vec3()
