+++ worldwind/src/commonMain/kotlin/earth/worldwind/util/kgl/Kgl.kt
@@ -96,7 +96,6 @@ const val GL_INVALID_FRAMEBUFFER_OPERATION = 0x506
 const val GL_CW = 0x0900
 const val GL_CCW = 0x0901
 const val GL_LINE_WIDTH = 0x0B21
-const val GL_POINT_SIZE_RANGE = 0x0B12
 const val GL_SMOOTH_POINT_SIZE_RANGE = 0x0B12
 const val GL_ALIASED_POINT_SIZE_RANGE = 0x846D
 const val GL_ALIASED_LINE_WIDTH_RANGE = 0x846E
